Separate app config form from app modal component
